Task-1-Python-Introduction/requirements.txt
annotated-types==0.5.0
mysql-connector-python==8.1.0
mysql-connector-python-rf==2.2.2
protobuf==4.21.12
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
typing_extensions==4.7.1
